Add a property to pass through the initial assembly name
Right now we don't get the command line string for options until the
full design time build has completed; however the project system can
still give us the evaluated string which is likely to be close
enough. This allows features that want to get symbol names for source
to have a better chance of having something that they can use and
look up in caches or in the cloud.