Fix faulty cast in state machine rewriter (#55063)
HoistRefInitialization assumed that the declarator syntax for locals would always be an await expression, but this is not true for switch expressions. CalculateLocalSyntaxOffset already correctly handles switch expressions, so we just needed to update this call site. I also inspected any other calls to CalculateLocalSyntaxOffset to make sure that nothing else was making such an assumption, but I didn't see any that are concerning.