Reduce array allocations when constructing options
Most `Option<T>`, `PerLanguageOption<T>`, `Option2<T>` and
`PerLanguageOption2<T>` instances are constructred by calling a constructor
that takes a params array. This array is them immediately converted to
an immutable array, resulting in two array allocations and a copy. This change
audits each option creation and changes the code to pass an immutable array,
avoiding an allocation and array copy per option.