Don't reference VS Shell binaries in EditorFeatures.Wpf
EditorFeatures.Wpf is supposed to be independent of the VS Shell, the
idea is this layer can be hosted in other applications using the
VS editor but not VS itself. This re-adds indirection for fetching
the dashboard colors, but now only does it when the dashboard is
opened.

While I was here I ended up renaming CodeAnalysisColors to
DashboardColors which is a bit clearer what it is.