Move compiler test ref assemblies to NuPkg (#55963)
* Move compiler test ref assemblies to NuPkg

The compiler and ide tests depend on in memory reference assemblies for constructing `Compilation` objects. These in memory reference assemblies are not offered by the platform and hence we were just generating them as a part of our build. That had a number of limitations though:

1. The reference assemblies for .NET Core are tied to the version of `Microsoft.NETCore.App.Ref.nupkg`. That is not a side effect free NuPkg. Referencing it has other impacts to the build. Particularly it is very difficult for us to build referencing multiple versions of that package hence we can't easily have tests that rely on a mix of `net5.0` and `net6.0`. That is very important for the compiler team though when working on cross cutting features.
2. The build infra puts all of the ref assemblies into a single project that is then referenced by every single test project we have. That means every test project carries a copy of the ref assemblies whether or not we use them. There is no realistic way to have a project reference only a subset of the ref assemblies. While the negative impact of build output size is often over stated this is one case where it has caused us real problems in the past (caused us to go over the 10gig limit for azure machines a few times). 

This change begins the process of moving us from generating the ref assemblies in our build to just referencing them as NuPkg from https://github.com/jaredpar/basic-reference-assemblies. The code for generating those assemblies is effectively what we had in our repo. I just cleaned it up a bit and changed it to generate assemblies per TFM so that it was easier to consume them on a "need" basis. Also because that build is much more disconnected it's easy to generate every TFM simultaneously and let us reference `net5.0` and `net6.0` at the same time.