Remove ErrorTypeException from native int decoder (#56350)
This change removes the `ErrorTypeException` type from
`NativeIntegerDecoder` and changes the code to propogate `null` + an
error state when it hits error types in decoding.

The reason for this change is to support VS 2022 performance
iniatitives. Even though the exception is thrown and handled on this
code path (never leaks) it was inhibbiting developers debugging with
first chance exceptions enabled as well as overall decreasing the
performance of this scenario. For context this exception is thrown three
times as much as the next common exception
(`OperationCanceledException`). Hence removing it

closes #49823