[Lightbulb Perf] Only force compute document diagnostics for the requested span
In the IDE, while determining in the background whether or not lightbulb should be shown for the current line, we only request compiler + analyzer diagnostics for the current line span.
Prior logic in CompilationWithAnalyzers was force computing compiler semantic diagnostics for the entire document, which not only meant unnecessary overhead for computing compiler diagnostics, but also generated compilation events for the entire document and hence an unnecessary overhead for computing analyzer diagnostics.

Verified that added unit test fails prior to this change.