Acquire the workspace lock asynchronously when applying batch updates
When the project system is loading a lot of projects at once, we can
end up with a bunch of threads all contending for the main
VisualStudioWorkspace lock, since we ultimately acquire the singleton
lock once we're updating CurrentSolution. This allows us to be async
so we aren't blocking up a bunch of thread pool threads.