Fix the assembly name used in generator identity
There's various places we need to determine a generator identity,
where we use the assembly name and generator type name. Incremental
source generators are "wrapped" in the Roslyn API in a regular
ISourceGenerator, and a special API gives you access to the underlying
object. We updated one of the functions we use for identity (getting
type name), but not the one for getting assembly name.

This could have been bad if you had two incremental generators with
the same type name in different assemblies, since we might consider
those the same.