Remove caching by assembly identity when using AssemblyLoadContexts
We had a cache in AnalyzerAssemblyLoader that tried to cache and reuse
the same assembly, keyed by assembly identity. This code makes some
sense to keep around for the .NET full framework case where we can't
load multiple assemblies with the same identity, but doesn't make
sense to keep around for our AssemblyLoadContext implementation.
The AssemblyLoadContext code correctly isolated dependencies and didn't
use this cache when resolving dependencies, but it was still active when
the public LoadFromPath entrypoint was called -- we'd hit the cache
there before doing anything else.

To ensure we don't have any identity-related caching logic reappearing
in the AssemblyLoadContext implementation, the identity-related logic
has been moved to DefaultAnalyzerAssemblyLoader.Desktop.cs, which
greatly simplifies the base type.