When we're loading dependencies, don't require exact versions
When we're loading analyzer dependencies, our current logic was to
require the exact version that was requested from the loading
assembly. This is problematic in a lot of cases where you would normally
have a binding redirect to ensure assemblies load cleanly in diamond
dependency scenarios. Without this, if you depend on B.dll and A.dll,
but B.dll depends on an older version of A.dll, you can't produce a
package that actually loads because there's no sane way to package both
A.dlls at the same time.

We avoid adding any logic here where we'll still try to use an older
version if that's still available. Consider a more complicated example
of:

1. Analyzer 1 depends on and ships B.dll (version 2) and A.dll
   (version 2); B.dll (version 2) depends on A.dll (version 1).
2. Analyzer 2 depends on A.dll (version 1) and ships that.

When loading Analyzer 1, we need a redirect to ensure the A that it
consumes and the A that B consumes are unified. We don't want the
presence of Analyzer 2 to change that. We could satisfy B.dll's request
with Ananlyer 2's dependency, but if Analyzer 1 expects type unification
then adding Analyzer 2 to the build would break it.

Fixes https://github.com/dotnet/roslyn/issues/53672