List-patterns: factor binding logic (#57318)
Co-authored-by: AlekseyTs <AlekseyTs@users.noreply.github.com>

This changes the bound tree representation of list patterns, slice patterns and implicit indexer accesses to avoid storing property or method symbols. Instead we store BoundIndexerAccess/BoundArrayAccess/BoundCall, which be obtain from existing "Bind*" methods and that we run through `CheckValue`. This ensures that we're not missing on all the validation rules which those include.
We intend to do some further refactoring, allowing `CheckValue` to properly deal with placeholders. This will allow us to store the Receiver separately from the IndexerOrSliceAccess and leverage existing placeholder infrastructure.