Ensure we don't silently capture any exceptions
In the synchronous dispose path, we were asserting that the task
completed; if it completed due to a fault, we were silently catching
that exception. This ensures we propagate the exception.

Under no "normal" circumstance we do we expect an exception to throw
when a batch is applied, so this is only needed for diagnosing faults.