Allow the FileCodeModel.Parent to be null
This returns a ProjectItem from the project system for a file, but in
the case of source generated files we won't be able to do that since
they're not represented by anything the project system knows about.

Coincidently, our unit tests always have been passing null the whole
time so this is something more or less we need to deal with regardless.