Fix a version conflict of System.Drawing.Common versions
It appears that some dependency of our MSBuild.UnitTests.csproj
project is bringing in 5.0.0 of System.Drawing.Common, while the
6.0 SDK is referencing binaries that reference 6.0.0 of
System.Drawing.Common, without carrying a proper reference to the 6.0.0
NuGet package. MSBuild can't resolve that, but adding an explicit
reference to 6.0.0 of System.Drawing.Common fixes that.

Upgrading to System.Drawing.Common then causes the platform analyzers
to realize that we're calling a Windows-only API on a cross-platform
DLL. We use HasValidFonts to work around a bug in some of our old CI
infrastructure, but the only place it's already used is already
restricted to Windows only, so we can just duplicate that check to
make everything happy again.