Delete Microsoft.CodeAnalysis.Workspaces.Desktop.UnitTests
The only code in here is testing CommandLineProject, but that was
forwarded into Microsoft.CodeAnalysis.Workspaces awhile ago. So
after moving these tests, the project is empty.

The only thing that could be tested in
Microsoft.CodeAnalysis.Workspaces.Desktop is our MEFv1 host services,
but we already have enough implicit testing on that as it is. (Famous
last words.)